PR fixes: updated Javadocs, removed wildcard imports, corrected names of ReturnsRxNTypes
