+++ rxjava2/src/main/kotlin/nz/co/kiwiandroiddev/mockito/rxjava/verification/rxjava2/ReturnsTrackedRx2Types.kt
@@ -1,36 +1,38 @@
-package nz.co.kiwiandroiddev.mockito.rxjava.verification.rxjava1
+package nz.co.kiwiandroiddev.mockito.rxjava.verification.rxjava2
 
+import io.reactivex.Completable
+import io.reactivex.Flowable
+import io.reactivex.Maybe
+import io.reactivex.Observable
+import io.reactivex.Single
 import nz.co.kiwiandroiddev.mockito.rxjava.verification.impl.SubscriptionOnDefaultMockedObservableEvent
 import org.mockito.internal.stubbing.defaultanswers.ReturnsEmptyValues
 import org.mockito.internal.util.MockUtil.getMockHandler
 import org.mockito.invocation.InvocationOnMock
-import rx.Completable
-import rx.Observable
-import rx.Single
 
 /**
- * Extends {@code org.mockito.internal.stubbing.defaultanswers.ReturnEmptyValues} to return a
- * default observable for observable-returning methods.
+ * Extends {@code org.mockito.internal.stubbing.defaultanswers.ReturnEmptyValues} to return default
+ * RxJava2 reactive types (e.g. Observable, Single, Maybe) from mock methods whose subscription events can be verified.
  *
- * This default observable enables two things:
+ * More specifically, this enables two things:
  *
- * - The number of subscriptions to this default observable can be verified with
+ * - The number of subscriptions on these default reactive objects can be verified with
  * verify(mock, wasSubscribedTo())
  *
- * - The observable will invokes an {@link Observer}'s {@link Observer#onError onError} method
- * when the Observer subscribes to it, reminding you that the method hasn't been stubbed yet
+ * - The reactive object will invoke its Observer's {@link Observer#onError onError} method
+ * on subscription, reminding you that the method hasn't been stubbed yet
  * (cf. a NPE would usually be thrown for unstubbed observable-returning methods when
  * simply using ReturnsEmptyValues)
  *
  * To use:
  *
- * For individual mocks, create the mock with Mockito.mock(Class::class.java, ReturnsTrackedRx1Observables())
+ * For individual mocks, create the mock with Mockito.mock(Class::class.java, ReturnsTrackedRx2Types())
  *
  * To use it with all mocks by default (i.e. globally), create a subclass of
  * DefaultMockitoConfiguration in package org.mockito.configuration and override its
  * getDefaultAnswer() method to return an instance of this class.
  */
-class ReturnsTrackedRx1Observables : ReturnsEmptyValues() {
+class ReturnsTrackedRx2Types : ReturnsEmptyValues() {
 
     override fun answer(invocation: InvocationOnMock): Any? {
         val methodReturnType = invocation.method.returnType
@@ -38,6 +40,8 @@ class ReturnsTrackedRx1Observables : ReturnsEmptyValues() {
             methodReturnType.isAssignableFrom(Observable::class.java) -> trackedObservable(invocation)
             methodReturnType.isAssignableFrom(Completable::class.java) -> trackedCompletable(invocation)
             methodReturnType.isAssignableFrom(Single::class.java) -> trackedSingle(invocation)
+            methodReturnType.isAssignableFrom(Maybe::class.java) -> trackedMaybe(invocation)
+            methodReturnType.isAssignableFrom(Flowable::class.java) -> trackedFlowable(invocation)
             else -> super.answer(invocation)
         }
     }
@@ -47,6 +51,13 @@ class ReturnsTrackedRx1Observables : ReturnsEmptyValues() {
                 recordSubscriptionOnMock(realMethodInvocation = invocation)
 
                 Observable.error<Any>(RuntimeException("missing stub observable for invocation: $invocation"))
+            }
+
+    private fun trackedFlowable(invocation: InvocationOnMock): Flowable<Any> =
+            Flowable.defer {
+                recordSubscriptionOnMock(realMethodInvocation = invocation)
+
+                Flowable.error<Any>(RuntimeException("missing stub flowable for invocation: $invocation"))
             }
 
     private fun trackedCompletable(invocation: InvocationOnMock): Completable =
@@ -61,6 +72,13 @@ class ReturnsTrackedRx1Observables : ReturnsEmptyValues() {
                 recordSubscriptionOnMock(realMethodInvocation = invocation)
 
                 Single.error<Any>(RuntimeException("missing stub single for invocation: $invocation"))
+            }
+
+    private fun trackedMaybe(invocation: InvocationOnMock): Maybe<Any> =
+            Maybe.defer {
+                recordSubscriptionOnMock(realMethodInvocation = invocation)
+
+                Maybe.error<Any>(RuntimeException("missing stub maybe for invocation: $invocation"))
             }
 
     private fun recordSubscriptionOnMock(realMethodInvocation: InvocationOnMock) {
